Moved text functions to new file
